Ansible/master_software/api/app/requirements.txt
fastapi==0.74.1
mysql-connector-python==8.0.28
pydantic==1.9.0
uvicorn==0.11.3

